Refactor: 게시글 상세 페이지 컴포넌트 분리
