+++ src/pages/articleDetail/components/ArticleTitle.tsx
@@ -1,0 +1,5 @@
+const ArticleTitle = () => {
+  return <div>ArticleTitle</div>;
+};
+
+export default ArticleTitle;
